Fixed problem with mini parser.
Parser would not work if client buffer was smaller than msg.
There is still a problem where if the client buffer is smaller than the msg, only the part of msg that fits inside the buffer will be sent. Since this is more of a proof of concept than a real parser, this problem will not be fixed.